refactor: change event.Version from int to uint64
Negative versions were never valid in event sourcing. The uint64
type makes impossible states unrepresentable at the type level.
Updates ParseVersion, Version methods, fuzz tests, and removes
dead negative-version-rejection tests (now enforced by the type
system itself).

+++ event/parser_fuzz_test.go
@@ -121,7 +121,7 @@ func FuzzVersion_Arithmetic(f *testing.F) {
 			return
 		}
 
-		v, err := event.ParseVersion(base)
+		v, err := event.ParseVersion(uint64(base))
 		if err != nil {
 			t.Fatalf("ParseVersion: %v", err)
 		}
@@ -206,7 +206,7 @@ func FuzzCheckVersionConflict(f *testing.F) {
 			return
 		}
 
-		v, err := event.ParseVersion(expected)
+		v, err := event.ParseVersion(uint64(expected))
 		if err != nil {
 			return
 		}
